requirements.txt
Adafruit-Blinka==8.46.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==4.1.4
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.72.1
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
load-dotenv==0.1.0
paho-mqtt==2.1.0
pyftdi==0.55.4
pyserial==3.5
python-dotenv==1.0.1
pyusb==1.2.1
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
sysv_ipc==1.1.0
typing_extensions==4.12.2
